Ask which ROS distribution to use when multiple ws are available
... and none is explicitly configured yet.

`/opt/ros/<ws>/` can be different from `$ROS_DISTRO` when system
maintainers prepare a separate workspace on top of a distribution.
Without this change there is no way to keep `rosdep` et al. happy with
knowing the correct distribution name and source the correct system
workspace nonetheless.

So when
```
$ ls /opt/ros/
/opt/ros/kinetic
/opt/ros/company
```
this patch will now ask on startup which one should be used for the workspace.